+++ HISTORY.rst
@@ -11,6 +11,7 @@ History
 * !142: Implemented processing of TESTFLAGS-VNIR and TESTFLAGS-SWIR layers in Level-1B reader.
 * !143: Updated copyright.
 * !147: Replaced the user-provided average_elevation by the mean ground elevation provided in the EnMAP metadata.
+* !140: Implemented automatic Copernicus DEM download to ensure high L2A quality, also if users do not provide a DEM.
 
 
 1.2.1 (2026-02-25)